Refatoring
