--- login bd/bancoAps.sql
@@ -1,168 +1,0 @@
--- MySQL dump 10.13  Distrib 5.7.24, for Win64 (x86_64)
---
--- Host: localhost    Database: apsjob
--- ------------------------------------------------------
--- Server version	5.7.24
-
-/*!40101 SET @OLD_CHARACTER_SET_CLIENT=@@CHARACTER_SET_CLIENT */;
-/*!40101 SET @OLD_CHARACTER_SET_RESULTS=@@CHARACTER_SET_RESULTS */;
-/*!40101 SET @OLD_COLLATION_CONNECTION=@@COLLATION_CONNECTION */;
-/*!40101 SET NAMES utf8 */;
-/*!40103 SET @OLD_TIME_ZONE=@@TIME_ZONE */;
-/*!40103 SET TIME_ZONE='+00:00' */;
-/*!40014 SET @OLD_UNIQUE_CHECKS=@@UNIQUE_CHECKS, UNIQUE_CHECKS=0 */;
-/*!40014 SET @OLD_FOREIGN_KEY_CHECKS=@@FOREIGN_KEY_CHECKS, FOREIGN_KEY_CHECKS=0 */;
-/*!40101 SET @OLD_SQL_MODE=@@SQL_MODE, SQL_MODE='NO_AUTO_VALUE_ON_ZERO' */;
-/*!40111 SET @OLD_SQL_NOTES=@@SQL_NOTES, SQL_NOTES=0 */;
-
---
--- Table structure for table `aluno`
---
-
-DROP TABLE IF EXISTS `aluno`;
-/*!40101 SET @saved_cs_client     = @@character_set_client */;
-/*!40101 SET character_set_client = utf8 */;
-CREATE TABLE `aluno` (
-  `idAluno` int(11) NOT NULL AUTO_INCREMENT,
-  `nome` varchar(45) DEFAULT NULL,
-  `email` varchar(45) DEFAULT NULL,
-  `instituicao` varchar(45) DEFAULT NULL,
-  `matricula` int(11) DEFAULT NULL,
-  PRIMARY KEY (`idAluno`)
-) ENGINE=InnoDB DEFAULT CHARSET=latin1;
-/*!40101 SET character_set_client = @saved_cs_client */;
-
---
--- Dumping data for table `aluno`
---
-
-LOCK TABLES `aluno` WRITE;
-/*!40000 ALTER TABLE `aluno` DISABLE KEYS */;
-/*!40000 ALTER TABLE `aluno` ENABLE KEYS */;
-UNLOCK TABLES;
-
---
--- Table structure for table `artigo`
---
-
-DROP TABLE IF EXISTS `artigo`;
-/*!40101 SET @saved_cs_client     = @@character_set_client */;
-/*!40101 SET character_set_client = utf8 */;
-CREATE TABLE `artigo` (
-  `idArtigo` int(11) NOT NULL AUTO_INCREMENT,
-  `titulo` varchar(45) DEFAULT NULL,
-  `autores` varchar(45) DEFAULT NULL,
-  `resumo` text,
-  `arquivo` varchar(45) DEFAULT NULL,
-  `status` varchar(45) DEFAULT NULL,
-  PRIMARY KEY (`idArtigo`)
-) ENGINE=InnoDB DEFAULT CHARSET=latin1;
-/*!40101 SET character_set_client = @saved_cs_client */;
-
---
--- Dumping data for table `artigo`
---
-
-LOCK TABLES `artigo` WRITE;
-/*!40000 ALTER TABLE `artigo` DISABLE KEYS */;
-/*!40000 ALTER TABLE `artigo` ENABLE KEYS */;
-UNLOCK TABLES;
-
---
--- Table structure for table `eventos`
---
-
-DROP TABLE IF EXISTS `eventos`;
-/*!40101 SET @saved_cs_client     = @@character_set_client */;
-/*!40101 SET character_set_client = utf8 */;
-CREATE TABLE `eventos` (
-  `ev_id` int(11) NOT NULL AUTO_INCREMENT,
-  `nome_ev` varchar(45) DEFAULT NULL,
-  `sigla` varchar(45) DEFAULT NULL,
-  `data_in` date DEFAULT NULL,
-  `data_fn` date DEFAULT NULL,
-  `data_sub_in` date DEFAULT NULL,
-  `data_sub_fn` date DEFAULT NULL,
-  `palavra_chave` varchar(250) DEFAULT NULL,
-  `area_conc` varchar(45) DEFAULT NULL,
-  `situacao` varchar(45) DEFAULT NULL,
-  PRIMARY KEY (`ev_id`)
-) ENGINE=InnoDB DEFAULT CHARSET=latin1;
-/*!40101 SET character_set_client = @saved_cs_client */;
-
---
--- Dumping data for table `eventos`
---
-
-LOCK TABLES `eventos` WRITE;
-/*!40000 ALTER TABLE `eventos` DISABLE KEYS */;
-/*!40000 ALTER TABLE `eventos` ENABLE KEYS */;
-UNLOCK TABLES;
-
---
--- Table structure for table `login`
---
-
-DROP TABLE IF EXISTS `login`;
-/*!40101 SET @saved_cs_client     = @@character_set_client */;
-/*!40101 SET character_set_client = utf8 */;
-CREATE TABLE `login` (
-  `usuario` varchar(45) NOT NULL,
-  `senha` varchar(45) DEFAULT NULL,
-  `tipo_user` varchar(45) DEFAULT NULL,
-  PRIMARY KEY (`usuario`),
-  UNIQUE KEY `senha_UNIQUE` (`senha`)
-) ENGINE=MyISAM DEFAULT CHARSET=latin1;
-/*!40101 SET character_set_client = @saved_cs_client */;
-
---
--- Dumping data for table `login`
---
-
-LOCK TABLES `login` WRITE;
-/*!40000 ALTER TABLE `login` DISABLE KEYS */;
-INSERT INTO `login` VALUES ('jv','pirocadegato','aluno');
-/*!40000 ALTER TABLE `login` ENABLE KEYS */;
-UNLOCK TABLES;
-
---
--- Table structure for table `professor`
---
-
-DROP TABLE IF EXISTS `professor`;
-/*!40101 SET @saved_cs_client     = @@character_set_client */;
-/*!40101 SET character_set_client = utf8 */;
-CREATE TABLE `professor` (
-  `idProfessor` int(11) NOT NULL AUTO_INCREMENT,
-  `nome` varchar(45) DEFAULT NULL,
-  `email` varchar(45) DEFAULT NULL,
-  `instituicao` varchar(45) DEFAULT NULL,
-  `titulacao` varchar(45) DEFAULT NULL,
-  `area_pesq` varchar(45) DEFAULT NULL,
-  PRIMARY KEY (`idProfessor`)
-) ENGINE=InnoDB DEFAULT CHARSET=latin1;
-/*!40101 SET character_set_client = @saved_cs_client */;
-
---
--- Dumping data for table `professor`
---
-
-LOCK TABLES `professor` WRITE;
-/*!40000 ALTER TABLE `professor` DISABLE KEYS */;
-/*!40000 ALTER TABLE `professor` ENABLE KEYS */;
-UNLOCK TABLES;
-
---
--- Dumping events for database 'apsjob'
---
-/*!40103 SET TIME_ZONE=@OLD_TIME_ZONE */;
-
-/*!40101 SET SQL_MODE=@OLD_SQL_MODE */;
-/*!40014 SET FOREIGN_KEY_CHECKS=@OLD_FOREIGN_KEY_CHECKS */;
-/*!40014 SET UNIQUE_CHECKS=@OLD_UNIQUE_CHECKS */;
-/*!40101 SET CHARACTER_SET_CLIENT=@OLD_CHARACTER_SET_CLIENT */;
-/*!40101 SET CHARACTER_SET_RESULTS=@OLD_CHARACTER_SET_RESULTS */;
-/*!40101 SET COLLATION_CONNECTION=@OLD_COLLATION_CONNECTION */;
-/*!40111 SET SQL_NOTES=@OLD_SQL_NOTES */;
-
--- Dump completed on 2019-09-14 20:40:00
