--- login bd/instruções.txt
@@ -1,10 +1,0 @@
-instale o node js
-cmd
-Express - Instale com o comando: npm install express .
-Sessões Expressas - Instale com o comando: npm install express-session .
-MySQL para Node.js - Instale com o comando: npm install mysql .
-
-carregar o arquivo bancoAps.sql no mysql na parte de dataimport
-
-depois so rodar o app.js no node :D
-
